move fetcher and scroll to common lib
